dnstask: move resolving responsibilities to module
The main process (xping.c) only knows the hostname of the target line
by the way of the module. By making it the modules responsibility to
resolve the hostname, sharing the struct target may not be nessecary in
the future.

This rearranges the was address family reporting works. Now each module
will have to call back with the address family (since they own the
resolving talk). This is a bit kludgy.